Upstreamed dreamer changes
