+++ torchrl/modules/tensordict_module/__init__.py
@@ -8,3 +8,4 @@ from .common import *
 from .exploration import *
 from .sequence import *
 from .probabilistic import *
+from .world_models import *
